% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muc.R
\name{summary.muc}
\alias{summary.muc}
\alias{print.muc}
\title{Summarizing output from \code{\link{muc}}}
\usage{
\method{summary}{muc}(object, ...)

\method{print}{muc}(x, ...)
}
\arguments{
\item{object}{an object of class `\code{muc}`.}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{an object of class \code{summary.muc}.}
}
\description{
This function is a method for class \code{\link{muc}} objects.
}
\examples{
#For examples see example(muc)

}
\seealso{
\code{\link{muc}}
}
